requirements.txt
flask
docker
python_on_whales
pyyaml